backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
authlib==1.3.1
httpx==0.27.2
python-jose[cryptography]==3.3.0
itsdangerous==2.2.0
SQLAlchemy==2.0.36
alembic==1.13.3
psycopg[binary]==3.2.3
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
email-validator==2.2.0
